cnn_vit_retrieval/requirements.txt
torch
torchvision
faiss-cpu
pdf2image
pillow
opencv-python
pytesseract
numpy
tqdm
